refactor react setup
